Backbone conversion 1: Instance properties
This is the first part of converting a Plain JavaScript "class" to a
Backbone Model.

With this change we are creating a new plain JS object called
`instanceProps` and putting each instance method inside of it as a
property. Previously we were putting each instance method on the
`Game.prototype` object, which will no longer exist after the conversion
is complete.

NOTE: Test will not be working after this change.