Simple Backbone app using Game model
This is mostly to verify that the Backbone Model is working
appropriately and can be used in a Backbone app.

The view layer is just displaying the Tic-Tac-Toe board contents
directly from the model, rather than building an HTML view from it.

There is not yet any indication of win status, or a way to easily
restart the game (other than refreshing the page).